% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_enr.R
\name{call_enr}
\alias{call_enr}
\title{Internal function \code{call_enr}}
\usage{
call_enr(clusters)
}
\arguments{
\item{clusters}{A subnetwork clustered using edge betweenness algorithm of \pkg{igraph}.}
}
\description{
This function is internally used to perform enrichment analysis employing ENRICHR API.
}
\author{
Murodzhon Akhmedov
}
\keyword{internal}

