% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCSF_phos.R
\name{PCSF_phos}
\alias{PCSF_phos}
\title{Prize-collecting Steiner Forest (PCSF) with phospho data}
\usage{
PCSF_phos(
  ppi,
  terminals,
  ksi,
  phosterminals,
  n = 100,
  r = 0.2,
  w = 4,
  b = 2,
  mu = 4e-04,
  dummies = c(),
  fname
)
}
\arguments{
\item{ppi}{An interaction network as data.frame with a 'from' and 'to' and 'cost' column}

\item{terminals}{A list of terminal genes with prizes to be analyzed in the PCSF context. 
A named \code{numeric} vector, where terminal genes are named same as in the interaction network 
and numeric values correspond to the importance of the gene within the study.}

\item{ksi}{A kinase-substrate interaction network as a data from with a 'from' and 'to' and 'cost' column.}

\item{phosterminals}{A list of terminal phosphosites with prizes with the format 'GENE-[S|T|YPosition]'}

\item{n}{An \code{integer} value to determine the number of runs with random noise added edge costs. 
A default value is 100.}

\item{r}{A \code{numeric} value to determine additional random noise to edge costs. 
A random noise upto r percent of the edge cost is added to each edge. A default value is 0.1}

\item{w}{A  \code{numeric} value for tuning the number of trees in the output. A default value is 2.}

\item{b}{A  \code{numeric} value for tuning the node prizes. A default value is 1.}

\item{mu}{A  \code{numeric} value for a hub penalization. A default value is 0.0005.}

\item{dummies}{A list of nodes that are to connected to the root of the tree. If missing the root will be connected to all terminals.}
}
\value{
The final subnetwork obtained by taking the union of the PCSF outputs generated by 
adding a random noise to edge costs each time. It returns an \pkg{igraph} object with the node prize 
and edge cost attributes representing the total number of show ups throughout all runs.
}
\description{
\code{PCSF_phos} returns a network with extensions provided for kinase-substrate interactions
}
\details{
In order to increase the robustness of the resulting structure, 
it is recommended to solve the PCSF several times on the same network
while adding some noise to the edge costs each time, and combine all results 
in a final subnetwork. The union of all outputs may explain 
the underlying biology better.
}
\examples{
\dontrun{
library("PCSF")
data("STRINGv12")
data("PSP_NK")
data("Tgfb_phospho")
data("Tgfb_psites")
terminals <-Tgfb_phospho
phosterminals <- Tgfb_psites
subnet <- PCSF_phos(STRINGv12, terminals, PSP_NK, phosterminals,n = 10, r =0.1, w = 2, b = 2, mu = 0.0005,fname='phos_test')
}

}
\references{
Akhmedov M., LeNail A., Bertoni F., Kwee I., Fraenkel E., and Montemanni R. (2017)
A Fast Prize-Collecting Steiner Forest Algorithm for Functional Analyses in Biological Networks.
\emph{Lecture Notes in Computer Science}, to appear.
}
\seealso{
\code{\link{PCSF}}, \code{\link{plot.PCSFe}}
}
\author{
Sara gosline
}
