% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_ks_interactome.R
\docType{data}
\name{PSP_NK}
\alias{PSP_NK}
\title{Interactome from phosphosite plus and networkin combined
 
An interactome data set in which the nodes are named with gene symbols. This is now updated for v12 in 2024}
\format{
A data frame with proteins connected to their sites and sites connected to the substrate protein
}
\source{
Previously collected list of kinases: https://raw.githubusercontent.com/PNNL-CompBio/amlresistancenetworks/refs/heads/master/inst/PSP%26NetworKIN_Kinase_Substrate_Dataset_July2016.csv
}
\usage{
PSP_NK
}
\description{
Interactome from phosphosite plus and networkin combined
 
An interactome data set in which the nodes are named with gene symbols. This is now updated for v12 in 2024
}
\keyword{data}
