% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_interactome.R
\docType{data}
\name{STRING}
\alias{STRING}
\title{Protein-protein interaction network data}
\format{A data frame with three variables, where each row corresponds to 
an edge in which the first element is a \code{head}, the second 
element is a \code{tail}, and the last element represents the \code{cost} of the edge.}
\source{
iref_mitab_miscore_2013_08_12_interactome.txt \url{https://github.com/fraenkel-lab/OmicsIntegrator/tree/master/data}
}
\usage{
STRING
}
\description{
An interactome data set in which the nodes are named with gene symbols
}
\keyword{data}
