% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_interactomev12.R
\docType{data}
\name{STRINGv12}
\alias{STRINGv12}
\title{Protein-protein interaction network data}
\format{
A data frame with three variables, where each row corresponds to 
an edge in which the first element is a \code{head}, the second 
element is a \code{tail}, and the last element represents the \code{cost} of the edge.
}
\source{
STRING \url{https://stringdb-downloads.org/download/protein.physical.links.v12.0/9606.protein.physical.links.v12.0.txt.gz}
}
\usage{
STRINGv12
}
\description{
An interactome data set in which the nodes are named with gene symbols. This is now updated for v12 in 2024
}
\keyword{data}
