% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_phosphoproteomics.R
\docType{data}
\name{Tgfb_phospho}
\alias{Tgfb_phospho}
\title{Phosphoproteomic data}
\format{A named \code{numeric} vector, where terminal genes are named same as 
in the interaction network and numeric values correspond to the importance of 
the gene within the study.}
\source{
Tgfb_phos.txt \url{https://github.com/fraenkel-lab/OmicsIntegrator/tree/master/example/a549}
}
\usage{
Tgfb_phospho
}
\description{
This dataset contains differential phosphoproteomic data derived from H358 cells, 
a model of lung cancer, that were stimulated with TGF-b.
}
\keyword{data}
