% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_ks_interactome.R, R/data_psites.R
\docType{data}
\name{Tgfb_psites}
\alias{Tgfb_psites}
\title{Random phosphosite weights}
\format{
A list of 100 weights representing phosphosites in the kinase-substrate network

A named \code{numeric} vector, where terminal genes are named same as 
in the interaction network and numeric values correspond to the importance of 
the gene within the study.
}
\source{
Randomly generated weights of phosphosites

randomly generrated to test phos function \url{https://github.com/fraenkel-lab/OmicsIntegrator/tree/master/example/a549}
}
\usage{
Tgfb_psites

Tgfb_psites
}
\description{
An interactome data set in which the nodes are named with gene symbols. This is now updated for v12 in 2024

This dataset contains differential phosphoproteomic data derived from H358 cells, 
a model of lung cancer, that were stimulated with TGF-b.
}
\keyword{data}
