% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_psites.R
\docType{data}
\name{Tgfb_psites}
\alias{Tgfb_psites}
\title{Randomly generated phosphosites from tgfb experiment}
\format{
A named \code{numeric} vector, where terminal genes are named same as 
in the interaction network and numeric values correspond to the importance of 
the gene within the study.
}
\source{
randomly generrated to test phos function \url{https://github.com/fraenkel-lab/OmicsIntegrator/tree/master/example/a549}
}
\usage{
Tgfb_psites
}
\description{
This dataset contains differential phosphoproteomic data derived from H358 cells, 
a model of lung cancer, that were stimulated with TGF-b.
}
\keyword{data}
