% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_ks_interactome.R
\docType{data}
\name{Tgfb_psites}
\alias{Tgfb_psites}
\title{Random phosphosite weights}
\format{
A list of 100 weights representing phosphosites in the kinase-substrate network
}
\source{
Randomly generated weights of phosphosites
}
\usage{
Tgfb_psites
}
\description{
An interactome data set in which the nodes are named with gene symbols. This is now updated for v12 in 2024
}
\keyword{data}
